backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
requests
pytesseract
Pillow
